Python Proj/tesser/requirements.txt
pdf2image
pytesseract
Pillow
Flask